# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DesiFinds API - Indian product recommendation engine
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SearchInput {
  query: string;
  /** @nullable */
  apiKey?: string | null;
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface ChatInput {
  message: string;
  history?: ChatMessage[];
  /** @nullable */
  apiKey?: string | null;
}

export interface Product {
  id: string;
  brand: string;
  name: string;
  description: string;
  category: string;
  tags: string[];
  materials: string[];
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  rating: number;
  reviewCount: number;
  productUrl: string;
  imageUrl: string;
  brandUrl: string;
  reviewSummary: string;
  madeInIndia: boolean;
  badges: string[];
}

export interface ChatResult {
  response: string;
  history: ChatMessage[];
  retrievedProducts: Product[];
}

export interface IngestInput {
  apiKey: string;
  force?: boolean;
}

export interface ProductMatch {
  product: Product;
  matchScore: number;
  matchReason: string;
  valueProp: string;
  /** @nullable */
  priceSavings?: number | null;
}

export interface ProductAnalysis {
  category: string;
  features: string[];
  materials: string[];
  priceRange: string;
  aestheticStyle: string;
  originalBrand: string;
}

export interface WorkflowStep {
  name: string;
  status: string;
  output: string;
}

export interface SearchResult {
  query: string;
  analysis: ProductAnalysis;
  matches: ProductMatch[];
  workflow: WorkflowStep[];
  aiPowered: boolean;
}

export interface ProductListResult {
  products: Product[];
  total: number;
  offset: number;
  limit: number;
}

export interface CategoryStat {
  name: string;
  count: number;
  icon: string;
}

export interface Brand {
  name: string;
  slug: string;
  description: string;
  logoUrl: string;
  websiteUrl: string;
  categories: string[];
  founded: string;
  featured: boolean;
}

export interface TrendingResult {
  popularSearches: string[];
  topProducts: Product[];
  newArrivals: Product[];
}

export type TriggerIngestion200 = {
  status: string;
  message: string;
};

export type ListProductsParams = {
category?: string;
brand?: string;
minPrice?: number;
maxPrice?: number;
minRating?: number;
q?: string;
limit?: number;
offset?: number;
};