# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DesiFinds API - Indian product recommendation engine
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatInput {
  message: string;
  history?: ChatMessage[];
  /** @nullable */
  apiKey?: string | null;
}