# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DesiFinds API - Indian product recommendation engine
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  role: string;
  content: string;
}