# lib/api-zod/src/generated/types/chatResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DesiFinds API - Indian product recommendation engine
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';
import type { Product } from './product';

export interface ChatResult {
  response: string;
  history: ChatMessage[];
  retrievedProducts: Product[];
}