# lib/api-zod/src/generated/types/ingestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DesiFinds API - Indian product recommendation engine
 * OpenAPI spec version: 0.1.0
 */

export interface IngestInput {
  apiKey: string;
  force?: boolean;
}